feat: upgrade Material-UI from v5 to v7
Completed incremental migration from MUI v5.15.18 to v7.3.4 following the v5→v6→v7 path.

### Major Changes:
- Updated @mui/material, @mui/icons-material, @mui/lab to v7.3.4
- Fixed import paths: @mui/material/styles → @mui/material for createTheme, useTheme, styled
- Changed tsconfig moduleResolution from "Node" to "Bundler"
- Replaced deprecated createPalette with createTheme pattern
- Created themeAugmentation.ts with comprehensive type augmentations

### Breaking Changes:
- TypeScript module augmentations moved to separate file
- Theme creation interfaces updated for v7 compatibility

### Fixes:
- All TypeScript errors resolved (0 errors)
- Build and lint passing successfully
