+++ src/components/AgronodAsideCard/AgronodAsideCard.tsx
@@ -4,7 +4,7 @@ import {
   SxProps,
 } from "@mui/material";
 import useMediaQuery from "@mui/material/useMediaQuery";
-import { useTheme } from "@mui/material/styles";
+import { useTheme } from "@mui/material";
 
 type AgronodAsideCardBaseProps = Pick<MuiCardProps, "variant" | "sx">;
