+++ src/components/Theme/ThemeProvider.tsx
@@ -1,14 +1,17 @@
 import React from "react";
-import { createTheme } from "@mui/material/styles";
+import { createTheme } from "@mui/material";
 import { deepmerge } from "@mui/utils";
 import {
   ThemeProvider as MUIThemeProvider,
   useTheme as useMUITheme,
-} from "@mui/material/styles";
+} from "@mui/material";
 import CssBaseline from "@mui/material/CssBaseline";
 import { ThemeProvider as EmotionThemeProvider } from "@emotion/react";
 import { ThemeOptions } from "@mui/material";
 import { baseThemeOptions, baseTheme } from "./baseTheme";
+
+// Import type augmentations
+import "./themeAugmentation";
 
 interface ThemeProviderProps {
   options?: ThemeOptions;
